Add backtest to pipeline
