+++ etna/metrics/__init__.py
@@ -15,3 +15,4 @@ from etna.metrics.metrics import MSLE
 from etna.metrics.metrics import R2
 from etna.metrics.metrics import SMAPE
 from etna.metrics.metrics import MedAE
+from etna.metrics.utils import compute_metrics
