+++ tests/test_pipeline/test_pipeline.py
@@ -10,12 +10,10 @@ from etna.datasets import TSDataset
 from etna.metrics import MAE
 from etna.metrics import MSE
 from etna.metrics import SMAPE
+from etna.metrics import Metric
 from etna.metrics import MetricAggregationMode
 from etna.models import LinearPerSegmentModel
-from etna.models import ProphetModel
 from etna.pipeline import Pipeline
-from etna.pipeline.backtest_utils import generate_folds_datasets
-from etna.pipeline.backtest_utils import validate_backtest_dataset
 from etna.transforms import AddConstTransform
 from etna.transforms import DateFlagsTransform
 
@@ -23,9 +21,7 @@ DEFAULT_METRICS = [MAE(mode=MetricAggregationMode.per_segment)]
 
 
 def test_fit(example_tsds):
-    """
-    Test that Pipeline correctly transforms dataset on fit stage
-    """
+    """Test that Pipeline correctly transforms dataset on fit stage."""
     original_ts = deepcopy(example_tsds)
     model = LinearPerSegmentModel()
     transforms = [AddConstTransform(in_column="target", value=10, inplace=True), DateFlagsTransform()]
@@ -36,9 +32,7 @@ def test_fit(example_tsds):
 
 
 def test_forecast(example_tsds):
-    """
-    Test that the forecast from the Pipeline is correct
-    """
+    """Test that the forecast from the Pipeline is correct."""
     original_ts = deepcopy(example_tsds)
 
     model = LinearPerSegmentModel()
@@ -55,20 +49,26 @@ def test_forecast(example_tsds):
     assert np.all(forecast_pipeline.df.values == forecast_manual.df.values)
 
 
-def test_validate_backtest_dataset(imbalanced_tsdf: TSDataset):
-    """Check validate_backtest_dataset behavior in case of small dataframe that
+@pytest.mark.parametrize("n_folds", (0, -1))
+def test_invalid_n_folds(catboost_pipeline: Pipeline, n_folds: int, example_tsdf: TSDataset):
+    """Test Pipeline.backtest behavior in case of invalid n_folds."""
+    with pytest.raises(ValueError):
+        _ = catboost_pipeline.backtest(ts=example_tsdf, metrics=DEFAULT_METRICS, n_folds=n_folds)
+
+
+@pytest.mark.parametrize("metrics", ([], [MAE(mode=MetricAggregationMode.macro)]))
+def test_invalid_backtest_metrics(catboost_pipeline: Pipeline, metrics: List[Metric], example_tsdf: TSDataset):
+    """Test Pipeline.backtest behavior in case of invalid metrics."""
+    with pytest.raises(ValueError):
+        _ = catboost_pipeline.backtest(ts=example_tsdf, metrics=metrics, n_folds=2)
+
+
+def test_validate_backtest_dataset(catboost_pipeline_big: Pipeline, imbalanced_tsdf: TSDataset):
+    """Test Pipeline.backtest behavior in case of small dataframe that
     can't be divided to required number of splits.
     """
     with pytest.raises(ValueError):
-        validate_backtest_dataset(ts=imbalanced_tsdf, n_folds=3, horizon=24)
-
-
-def test_invalid_metrics(example_tsds):
-    """Check TimeSeriesCrossValidation behavior in case of invalid metrics"""
-    pipeline = Pipeline(model=ProphetModel(), horizon=14)
-    with pytest.raises(ValueError):
-        metrics = [MAE(mode=MetricAggregationMode.macro)]
-        pipeline.backtest(ts=example_tsds, metrics=metrics, n_folds=2)
+        _ = catboost_pipeline_big.backtest(ts=imbalanced_tsdf, n_folds=3, metrics=DEFAULT_METRICS)
 
 
 def test_generate_expandable_timeranges_days():
@@ -85,7 +85,7 @@ def test_generate_expandable_timeranges_days():
         (("2021-01-01", "2021-03-08"), ("2021-03-09", "2021-03-20")),
         (("2021-01-01", "2021-03-20"), ("2021-03-21", "2021-04-01")),
     )
-    for i, stage_dfs in enumerate(generate_folds_datasets(ts, n_folds=3, horizon=12, mode="expand")):
+    for i, stage_dfs in enumerate(Pipeline._generate_folds_datasets(ts, n_folds=3, horizon=12, mode="expand")):
         for stage_df, borders in zip(stage_dfs, true_borders[i]):
             assert stage_df.index.min() == datetime.strptime(borders[0], "%Y-%m-%d").date()
             assert stage_df.index.max() == datetime.strptime(borders[1], "%Y-%m-%d").date()
@@ -105,7 +105,7 @@ def test_generate_expandable_timeranges_hours():
         (("2020-01-01 00:00:00", "2020-01-31 00:00:00"), ("2020-01-31 01:00:00", "2020-01-31 12:00:00")),
         (("2020-01-01 00:00:00", "2020-01-31 12:00:00"), ("2020-01-31 13:00:00", "2020-02-01 00:00:00")),
     )
-    for i, stage_dfs in enumerate(generate_folds_datasets(ts, horizon=12, n_folds=3, mode="expand")):
+    for i, stage_dfs in enumerate(Pipeline._generate_folds_datasets(ts, horizon=12, n_folds=3, mode="expand")):
         for stage_df, borders in zip(stage_dfs, true_borders[i]):
             assert stage_df.index.min() == datetime.strptime(borders[0], "%Y-%m-%d %H:%M:%S").date()
             assert stage_df.index.max() == datetime.strptime(borders[1], "%Y-%m-%d %H:%M:%S").date()
@@ -125,7 +125,7 @@ def test_generate_constant_timeranges_days():
         (("2021-01-13", "2021-03-08"), ("2021-03-09", "2021-03-20")),
         (("2021-01-25", "2021-03-20"), ("2021-03-21", "2021-04-01")),
     )
-    for i, stage_dfs in enumerate(generate_folds_datasets(ts, horizon=12, n_folds=3, mode="constant")):
+    for i, stage_dfs in enumerate(Pipeline._generate_folds_datasets(ts, horizon=12, n_folds=3, mode="constant")):
         for stage_df, borders in zip(stage_dfs, true_borders[i]):
             assert stage_df.index.min() == datetime.strptime(borders[0], "%Y-%m-%d").date()
             assert stage_df.index.max() == datetime.strptime(borders[1], "%Y-%m-%d").date()
@@ -144,7 +144,7 @@ def test_generate_constant_timeranges_hours():
         (("2020-01-01 12:00:00", "2020-01-31 00:00:00"), ("2020-01-31 01:00:00", "2020-01-31 12:00:00")),
         (("2020-01-02 00:00:00", "2020-01-31 12:00:00"), ("2020-01-31 13:00:00", "2020-02-01 00:00:00")),
     )
-    for i, stage_dfs in enumerate(generate_folds_datasets(ts, horizon=12, n_folds=3, mode="constant")):
+    for i, stage_dfs in enumerate(Pipeline._generate_folds_datasets(ts, horizon=12, n_folds=3, mode="constant")):
         for stage_df, borders in zip(stage_dfs, true_borders[i]):
             assert stage_df.index.min() == datetime.strptime(borders[0], "%Y-%m-%d %H:%M:%S").date()
             assert stage_df.index.max() == datetime.strptime(borders[1], "%Y-%m-%d %H:%M:%S").date()
@@ -165,6 +165,7 @@ def test_get_metrics_interface(
 
 
 def test_get_forecasts_interface_daily(catboost_pipeline: Pipeline, big_daily_example_tsdf: TSDataset):
+    """Check that Pipeline.backtest returns forecasts in correct format."""
     _, forecast, _ = catboost_pipeline.backtest(ts=big_daily_example_tsdf, metrics=DEFAULT_METRICS)
     expected_columns = sorted(
         ["regressor_target_lag_10", "regressor_target_lag_11", "regressor_target_lag_12", "fold_number", "target"]
@@ -173,6 +174,7 @@ def test_get_forecasts_interface_daily(catboost_pipeline: Pipeline, big_daily_ex
 
 
 def test_get_forecasts_interface_hours(catboost_pipeline: Pipeline, example_tsdf: TSDataset):
+    """Check that Pipeline.backtest returns forecasts in correct format with non-daily seasonality."""
     _, forecast, _ = catboost_pipeline.backtest(ts=example_tsdf, metrics=DEFAULT_METRICS)
     expected_columns = sorted(
         ["regressor_target_lag_10", "regressor_target_lag_11", "regressor_target_lag_12", "fold_number", "target"]
@@ -181,12 +183,14 @@ def test_get_forecasts_interface_hours(catboost_pipeline: Pipeline, example_tsdf
 
 
 def test_get_fold_info_interface_daily(catboost_pipeline: Pipeline, big_daily_example_tsdf: TSDataset):
+    """Check that Pipeline.backtest returns info dataframe in correct format."""
     _, _, info_df = catboost_pipeline.backtest(ts=big_daily_example_tsdf, metrics=DEFAULT_METRICS)
     expected_columns = ["fold_number", "test_end_time", "test_start_time", "train_end_time", "train_start_time"]
     assert expected_columns == list(sorted(info_df.columns))
 
 
 def test_get_fold_info_interface_hours(catboost_pipeline: Pipeline, example_tsdf: TSDataset):
+    """Check that Pipeline.backtest returns info dataframe in correct format with non-daily seasonality."""
     _, _, info_df = catboost_pipeline.backtest(ts=example_tsdf, metrics=DEFAULT_METRICS)
     expected_columns = ["fold_number", "test_end_time", "test_start_time", "train_end_time", "train_start_time"]
     assert expected_columns == list(sorted(info_df.columns))
@@ -194,6 +198,7 @@ def test_get_fold_info_interface_hours(catboost_pipeline: Pipeline, example_tsdf
 
 @pytest.mark.long
 def test_backtest_with_n_jobs(catboost_pipeline: Pipeline, big_example_tsdf: TSDataset):
+    """Check that Pipeline.backtest gives the same results in case of single and multiple jobs modes."""
     ts1 = deepcopy(big_example_tsdf)
     ts2 = deepcopy(big_example_tsdf)
     pipeline_1 = deepcopy(catboost_pipeline)
